Remove value-level hash algorithms.
We can use NamedHashAlgorithm to get a name where relevant, and soon
the pair of (recursive, Proxy) :: (Bool, Proxy algo) will be replaced
with a separate type which will let us drop the proxies.

Also, we drop logic from addToStore in the remote impl that was copied
from C++ compatibility code for very old daemon versions.