hnix-store-remote: Remove erroneous path putting/getting.
mkPath's logic was completely wrong (tried to make a PathName out of
the entire path, made up a hash based on the whole path rather than
parsing base32) and putPath ignored the store directory and the hash
part.

Much of the code that depended on these functions was actually
correct, but ultimately did the wrong thing. We can resurrect those
from git once the primitives are correctly implemented.