Rewrite NarEffects in terms of RawFilePath.
Nar handling, and Nix in general, only really makes sense in the
context of Unix-style file paths. And even outside of Unix land,
FilePath is a bad interface.

This commit also moves narEffectsIO into a local definition in the
test suite, to avoid a unix dependency in the core library. This
should eventually be moved into a separate library (and defined
directly in terms of the RawFilePath operations).