treewide: migrate crypto(hash-* -> nite)
Reference:

Main cause:

https://github.com/haskell-hvr/cryptohash-sha512/issues/7

The whole `cryptohash-*` package family is abandoned, there is no
signs of maintainer activity there, so it stopped following Haskell ecosystem &
`base` releases. Knowing the human history & situation around it - it would not
be reviwed, which gives experience to not hardcode on the (specifically when
emotional) dependency.

Experience I drawn from this story is to keep things simplier when possible &
have more flexible systems as a result code.
It was "a bit too much" for what hashing is, for the code to have 2 hashing type
systems (external & internal) & reinventment of `HashAlgorithm` type duplicate.
The whole code was really rigid with a lot of type applicating the data kinds,
those are dependent type features & should be used cautiously, since interface
became rigid to changes, so afterwards it is easier & effective to dismantle
and recreate the subsystem then to evolve it.

Previous hashing history:
https://github.com/haskell-nix/hnix-store/issues/156
https://github.com/haskell-nix/hnix-store/issues/142
https://github.com/haskell-nix/hnix-store/pull/93
https://github.com/haskell-nix/hnix-store/issues/92
https://github.com/haskell-nix/hnix-store/issues/90
https://github.com/haskell-nix/hnix-store/issues/83
https://github.com/haskell-nix/hnix-store/pull/64
https://github.com/haskell-nix/hnix-store/pull/38
https://github.com/haskell-nix/hnix-store/pull/32
https://github.com/haskell-nix/hnix-store/pull/31
https://github.com/haskell-nix/hnix-store/pull/28
https://github.com/haskell-nix/hnix-store/pull/27
https://github.com/haskell-nix/hnix-store/pull/25
https://github.com/haskell-nix/hnix-store/issues/18
https://github.com/haskell-nix/hnix-store/pull/14