Add outputName and drvOutput type parameters to Derivation
Allows to use custom types for both sides of the map.

The main reason for this is `hnix-store` which now uses
`DSum HashAlgo Digest` which combines `hashAlgo` and `hash` fields
of the `DerivationOutput` and `OutputName` type for the keys,
which is similar to how the newtype is done in the mainline Nix.